+++ src/main/java/tn/esprit/bank/vo/TransactionPeriodicVO.java
@@ -4,13 +4,14 @@ import lombok.Data;
 import tn.esprit.bank.enumeration.Periodicity;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Date;
 
 @Data
 public class TransactionPeriodicVO {
 
-    private BigDecimal accountNumberFrom;
-    private BigDecimal accountNumberTo;
+    private BigInteger accountNumberFrom;
+    private BigInteger accountNumberTo;
     private BigDecimal amount;
     private Periodicity periodicity;
     private Date startDate;
